Data settings builder + simulated/real data settings classes
