+++ src/io/data_settings/DataSettings.h
@@ -1,0 +1,48 @@
+#pragma once
+
+typedef double ScalarType;
+
+#include <fstream>
+#include <iostream>
+#include <string>
+
+#include "tinyxml2.h"
+
+namespace io {
+
+class DataSettings {
+public:
+  ////////////////////////////////////////////////////////////////////////////////////////////////
+  /// Constructor(s) / Destructor :
+  ////////////////////////////////////////////////////////////////////////////////////////////////
+
+  DataSettings(const char *xml_file);
+
+  ~DataSettings(); 
+  
+  /// Assignment operator, private to prevent copy
+  DataSettings& operator=(const DataSettings&);
+  
+    /// Copy constructor, private to prevent copy
+  DataSettings(const DataSettings&);
+
+  ////////////////////////////////////////////////////////////////////////////////////////////////
+  /// Encapsulation method(s) :
+  ////////////////////////////////////////////////////////////////////////////////////////////////
+  
+  virtual bool IsReal() const { return is_data_real_; }
+
+  
+private:
+
+  /// Read real data if true
+  bool is_data_real_;
+  
+  ////////////////////////////////////////////////////////////////////////////////////////////////
+  /// Method(s) :
+  ////////////////////////////////////////////////////////////////////////////////////////////////
+  
+    
+};
+
+}
